feat: rc-input
