+++ package.json
@@ -44,15 +44,17 @@
   "dependencies": {
     "@babel/runtime": "^7.10.1",
     "classnames": "^2.2.1",
+    "rc-input": "^0.0.1-alpha.7",
     "rc-resize-observer": "^1.0.0",
     "rc-util": "^5.7.0",
     "shallowequal": "^1.1.0"
   },
   "devDependencies": {
+    "@testing-library/react": "^13.3.0",
     "@types/classnames": "^2.2.9",
     "@types/enzyme": "^3.10.10",
-    "@types/react": "^16.9.2",
-    "@types/react-dom": "^16.9.0",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "@types/shallowequal": "^1.1.1",
     "@umijs/fabric": "^2.0.8",
     "coveralls": "^3.0.6",
@@ -65,12 +67,13 @@
     "father": "^2.13.4",
     "gh-pages": "^3.1.0",
     "husky": "^4.2.5",
+    "jest": "^28.1.3",
     "less": "^3.10.3",
     "np": "^5.1.0",
     "prettier": "^2.0.5",
     "pretty-quick": "^2.0.1",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-test-renderer": "^16.0.0"
   },
   "peerDependencies": {
